Develop

+++ src/repository/repository_chunk_has_chunk_hash_sqlite.go
@@ -1,16 +1,23 @@
 package repository
 
-import "database/sql"
+import (
+	"database/sql"
+	"errors"
+)
 
 type RepositoryChunkHasChunkHashSqlite struct{}
 
 func (rchchs RepositoryChunkHasChunkHashSqlite) Create(chunkId int64, chunkHashId int64, tx *sql.Tx) error {
-	query := `INSERT INTO chunks_has_chunk_hashes (chunk_id, chunk_hash_id) VALUES (?, ?)`
-	_, err := tx.Exec(query, chunkId, chunkHashId)
+	dml := `INSERT INTO chunks_has_chunk_hashes (chunk_id, chunk_hash_id) VALUES (?, ?)`
+	_, err := tx.Exec(dml, chunkId, chunkHashId)
 	return err
 }
 
-func (rchchs RepositoryChunkHasChunkHashSqlite) RemoveAll(tx *sql.Tx) error {
+func (rchchs RepositoryChunkHasChunkHashSqlite) RemoveAllWithTransaction(tx *sql.Tx) error {
+	if tx == nil {
+		return errors.New("transaction not found")
+	}
+
 	_, err := tx.Exec(`DELETE FROM chunks_has_chunk_hashes`)
 	if err != nil {
 		return err
@@ -21,7 +28,20 @@ func (rchchs RepositoryChunkHasChunkHashSqlite) RemoveAll(tx *sql.Tx) error {
 }
 
 func (rchchs RepositoryChunkHasChunkHashSqlite) RemoveByChunkId(id int64, tx *sql.Tx) error {
-	_, err := tx.Exec(`DELETE FROM chunks_has_chunk_hashes WHERE chunk_id = ?`, id)
+	result, err := tx.Exec(`DELETE FROM chunks_has_chunk_hashes WHERE chunk_id = ?`, id)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return ErrorRecordNotFound
+	}
+
 	return err
 }
 
@@ -68,8 +88,8 @@ func (rchchs RepositoryChunkHasChunkHashSqlite) GetChunkHashIdsByChunkId(chunkId
 	}
 	defer db.Close()
 
-	query := `SELECT chunk_hash_id FROM chunks_has_chunk_hashes WHERE chunk_id = ?`
-	rows, err := db.Query(query, chunkId)
+	dml := `SELECT chunk_hash_id FROM chunks_has_chunk_hashes WHERE chunk_id = ?`
+	rows, err := db.Query(dml, chunkId)
 	if err == sql.ErrNoRows {
 		return ids, nil
 	}
